+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased [4.1.0-SNAPSHOT]
 ### Added
-  - apply to IDEA task (#91)
+  - apply to IDEA task (in the project) (#91)
+  - apply to IDEA task (global settings) (#91)
 ### Changed
   - ?
 ### Removed